Changed the main.py file to run the main program logic, which uses functions from other modules to run the main program logic.
