+++ YandexDiscordRPC/version.ini
@@ -1,2 +1,2 @@
 [Version]
-version = 1.1.1
+version = 1.5.0
